requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
flask-migrate==4.0.5
python-dotenv==1.0.0
pymysql==1.0.3
tzdata
cairosvg==2.8.2
pillow==12.1.0
